# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Task queue backed by Elasticsearch",
        "title": "Tasques API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/tasques": {
            "post": {
                "description": "Creates a new Task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Add a new Task",
                "operationId": "create-task",
                "parameters": [
                    {
                        "description": "The request body",
                        "name": "newTask",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.NewTask"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    }
                }
            }
        },
        "/tasques/claims": {
            "post": {
                "description": "Claims a number of existing Tasks.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Claims a number of Tasks",
                "operationId": "claim-tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Worker ID",
                        "name": "X-TASQUES-WORKER-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The request body",
                        "name": "claim",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.Claim"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/task.Task"
                            }
                        }
                    }
                }
            }
        },
        "/tasques/claims/{queue}/{id}": {
            "delete": {
                "description": "Unclaims a claimed Task.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Unclaims a Task",
                "operationId": "unclaim-existing-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The Queue of the Task",
                        "name": "queue",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The id of the Task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Worker ID",
                        "name": "X-TASQUES-WORKER-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    },
                    "400": {
                        "description": "The Task is not currently claimed",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    },
                    "403": {
                        "description": "Worker currently has not claimed the Task",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    },
                    "404": {
                        "description": "Task does not exist",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    }
                }
            }
        },
        "/tasques/done/{queue}/{id}": {
            "put": {
                "description": "Marks a claimed Task as done.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Mark Task as Done",
                "operationId": "mark-claimed-task-done",
                "parameters": [
                    {
                        "description": "The request body",
                        "name": "success",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.Success"
                        }
                    },
                    {
                        "type": "string",
                        "description": "The Queue of the Task",
                        "name": "queue",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The id of the Task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Worker ID",
                        "name": "X-TASQUES-WORKER-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    },
                    "400": {
                        "description": "The Task is not currently claimed",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    },
                    "403": {
                        "description": "Worker currently has not claimed the Task",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    },
                    "404": {
                        "description": "Task does not exist",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    }
                }
            }
        },
        "/tasques/failed/{queue}/{id}": {
            "put": {
                "description": "Marks a claimed Task as failed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Mark Task as Failed",
                "operationId": "mark-claimed-task-failed",
                "parameters": [
                    {
                        "description": "The request body",
                        "name": "failure",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.Failure"
                        }
                    },
                    {
                        "type": "string",
                        "description": "The Queue of the Task",
                        "name": "queue",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The id of the Task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Worker ID",
                        "name": "X-TASQUES-WORKER-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    },
                    "400": {
                        "description": "The Task is not currently claimed",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    },
                    "403": {
                        "description": "Worker currently has not claimed the Task",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    },
                    "404": {
                        "description": "Task does not exist",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    }
                }
            }
        },
        "/tasques/reports/{queue}/{id}": {
            "put": {
                "description": "Reports in on a claimed Task.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Reports on a Task",
                "operationId": "report-on-claimed-task",
                "parameters": [
                    {
                        "description": "The request body",
                        "name": "newReport",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.NewReport"
                        }
                    },
                    {
                        "type": "string",
                        "description": "The Queue of the Task",
                        "name": "queue",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The id of the Task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Worker ID",
                        "name": "X-TASQUES-WORKER-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    },
                    "400": {
                        "description": "The Task is not currently claimed",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    },
                    "403": {
                        "description": "Worker currently has not claimed the Task",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    },
                    "404": {
                        "description": "Task does not exist",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    }
                }
            }
        },
        "/tasques/{queue}/{id}": {
            "get": {
                "description": "Retrieves a persisted Task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get a Task",
                "operationId": "get-existing-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The Queue of the Task",
                        "name": "queue",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The id of the Task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    },
                    "404": {
                        "description": "Task does not exist",
                        "schema": {
                            "$ref": "#/definitions/common.Body"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.Body": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Something went wrong :("
                }
            }
        },
        "common.Metadata": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "modified_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "version": {
                    "type": "object",
                    "$ref": "#/definitions/common.Version"
                }
            }
        },
        "common.Version": {
            "type": "object",
            "properties": {
                "primary_term": {
                    "type": "integer"
                },
                "seq_num": {
                    "type": "integer"
                }
            }
        },
...
        "task.Claim": {
            "type": "object",
            "required": [
                "queues"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 1
                },
                "block_for": {
                    "type": "string",
                    "example": "1s"
                },
                "queues": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "run-later",
                        "resize-images"
                    ]
                }
            }
        },
...
        "task.Failure": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                }
            }
        },
        "task.JsonObj": {
            "type": "object",
            "additionalProperties": true
        },
        "task.LastClaimed": {
            "type": "object",
            "required": [
                "claimed_at",
                "times_out_at",
                "worker_id"
            ],
            "properties": {
                "claimed_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "last_report": {
                    "type": "object",
                    "$ref": "#/definitions/task.Report"
                },
                "result": {
                    "type": "object",
                    "$ref": "#/definitions/task.Result"
                },
                "times_out_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
        "task.NewReport": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/task.ReportedData"
                }
            }
        },
        "task.NewTask": {
            "type": "object",
            "required": [
                "kind",
                "queue"
            ],
            "properties": {
                "args": {
                    "type": "object"
                },
                "context": {
                    "type": "object"
                },
                "kind": {
                    "type": "string",
                    "example": "sayHello"
                },
                "priority": {
                    "type": "integer"
                },
                "processing_timeout": {
                    "type": "string",
                    "example": "30m"
                },
                "queue": {
                    "type": "string",
                    "example": "run-later"
                },
                "retry_times": {
                    "type": "integer",
                    "example": 10
                },
                "run_at": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "task.Report": {
            "type": "object",
            "required": [
                "at"
            ],
            "properties": {
                "at": {
                    "type": "string",
                    "format": "date-time"
                },
                "data": {
                    "type": "object"
                }
            }
        },
        "task.ReportedData": {
            "$ref": "#/definitions/task.JsonObj"
        },
        "task.Result": {
            "type": "object",
            "required": [
                "at"
            ],
            "properties": {
                "at": {
                    "type": "string",
                    "format": "date-time"
                },
                "failure": {
                    "description": "Results. Only one of the following will be filled in at a given time",
                    "type": "object"
                },
                "success": {
                    "type": "object"
                }
            }
        },
        "task.Success": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                }
            }
        },
        "task.Task": {
            "type": "object",
            "required": [
                "attempted",
                "id",
                "kind",
                "last_enqueued_at",
                "metadata",
                "priority",
                "processing_timeout",
                "queue",
                "retry_times",
                "run_at",
                "state"
            ],
            "properties": {
                "args": {
                    "type": "object"
                },
                "attempted": {
                    "type": "integer"
                },
                "context": {
                    "type": "object"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string",
                    "example": "sayHello"
                },
                "last_claimed": {
                    "type": "object",
                    "$ref": "#/definitions/task.LastClaimed"
                },
                "last_enqueued_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "metadata": {
                    "type": "object",
                    "$ref": "#/definitions/common.Metadata"
                },
                "priority": {
                    "type": "integer"
                },
                "processing_timeout": {
                    "type": "string",
                    "example": "30m"
                },
                "queue": {
                    "type": "string",
                    "example": "run-later"
                },
                "recurring_task_id": {
                    "description": "Only populated if this is a Task that was spawned/enqueued by a Recurring Task definition",
                    "type": "string"
                },
                "retry_times": {
                    "type": "integer",
                    "example": 10
                },
                "run_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "state": {
                    "type": "string",
                    "example": "queued"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}